Created some New* funcs for Listeners and TargetGroups.
Added .Hash() funcs that can be used to tell if a listener or TG has changed, this is helpful since the TG struct doesn't include the targets. We may not need it for the Listener long term.

Main thing here is that I am loading all of the LoadBalancer/TargetGroup/Listener/Rule objects from AWS and keeping them in the albIngress list. Then when receiving an OnUpdate I am fetching those objects from the lastIngress list and adding references to the AWS objects to our newly created objects. If we go down this road the modify funcs should just compare what was provided from k8s to what is in the AWS objects.

Still got a long way to go, but I wanted to get a copy off of my laptop before I fly to Quebec in the morning!