Select subnets for ALB  when annotation is absent
- When the alb.ingress.kubernetes.io/subnets annotation is not present
on an ingress resource, the alb-ingress-controller attempts to qualify
2+ subnets to deploy in.

- Subnets that qualify are based on the tags of the subnets themselves.
In order for a subnet to be used to deploy an ALB the following tag
criteria must be met.

  - kubernetes.io/cluster/$CLUSTER_NAME where $CLUSTER_NAME is the same
cluster name specified on the ingress controller. The value of this tag
must be 'shared'.

  - kubernetes.io/role/alb-ingress the value of this tag should be
empty.

- After subnets matching the above 2 tags have been located, they are
checked to ensure 2 or more are in unique AZs, otherwise the ALB will
not be created. If 2 subnets share the same AZ, only 1 of the 2 is used.

- Resolves https://github.com/coreos/alb-ingress-controller/issues/7