Always use local (to cluster) VPC ID for ALBs
Previously the VPC ID, and thus VPC ALBs would be created within, was
detected by looking up the VPC based on the subnets the controller
intended to use. This approach can create conflicts when the subnets are
looked up by name (or perhaps other tags). This can cause another VPC
with identical subnet names to be resolved.

This new approach resolves the VPC ID based on the current EC2 instance
the alb-ingress-controller is running on.