Try to prevent and don't count predictable AWS API errors
This inserts a delay after creating an ALB listener because sometimes
interacting with a listener directly after creating it will produce
"listener not found" errors.

It also adds a context.Context value that is passed to Describe calls
where it is a normal situation that they report a "not found" error if
the function is called prior to creating the ALB/target group.