Allow TargetGroup endpoints outside the ELB VPC (#1862)
* Allow TargetGroup endpoints outside the ELB VPC

To support the scenario where the ELB is deployed in a separate VPC to
the EKS cluster, we need to conditionally set the `AvailabilityZone`
field to `all` when the target is outside of the ELB VPC.

This commit addresses this by checking if the pod IP is found
in the ELB VPC's CIDR or a secondary CIDR; to improve performance and
reduce calls to the AWS API, VPC info is cached for a configurable
amount of time, defaulting to 5 minutes.

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>

* Make isELBV2TargetInELBVPC more efficient

Don't need to check vpc.CidrBlock as it is part of vpc.CidrBlockAssociationSet. isIPinCIDR is now also a bit more robust in checking if an IP is part of a CIDR.

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>

* Add vpcID to defaultResourceManager

We're already passing this into NewDefaultResourceManager so no need to also add it to NetworkingManager.

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>

* Change VPC cache duration to 10m

Also switched cache value to be time.Duration from int.

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>

* Whoops, we should return an error if found...

* Use apimachinery/pkg/util/cache for caching

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>

* remove the aws-vpc-cache-ttl flag
fix unit tests

Co-authored-by: M00nF1sh <yyyng@amazon.com>