Fix Dockerfile go dependency caching
The current Dockerfile uses a bind mount when downloading the go dependencies.
This causes the entire project folder to be added to the Docker context when
only the `go.mod` and `go.sum` files are needed.  This ends up breaking the
caching of the dependencies as any change to any file in the project will cause
the docker context to change.  Removing this bind mount allows the dependencies
to be correctly cached.