Update VPCES annotations to be alpha
These annotations are new and should be introduced under and alpha
flag instead of the beta one currently used for everything else.

There is no default functionality for changing the version of an
annotation but we can explicitly add the alpha annotation
and assume users will follow the docs and use the alpha annotation
despite the beta annotation still actually working.