update unit tests for ctrl client/fake >0.15
controller-runtime >=0.15 does not support creating (or adding the field via Update()) objects with a DeletionTimestamp. To work around this we add an annotation `unit-test/delete` to mark the ingresses that we want to test deletion. We check for this annotation and then call Delete(). This will set the DeletionTimestamp to the current date and time so we use IgnoreOtherFields to skip then comparing want to got.

relevant controller-runtime discussion/pr:

- https://github.com/kubernetes-sigs/controller-runtime/issues/2184#issuecomment-1433401752
- https://github.com/kubernetes-sigs/controller-runtime/pull/2316